Django API/API/requirements.txt
django
djangorestframework
environs
 